% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condemned.R
\name{load_condemned_props}
\alias{load_condemned_props}
\title{Load condemned properties data}
\usage{
load_condemned_props()
}
\value{
A tibble of condemned properties with formatted dates
}
\description{
Loads condemned properties data from WPRDC. Returns a tibble with
formatted date column.
}
\examples{
\dontrun{
condemned <- load_condemned_props()
}
}
