% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipalities.R
\name{normalize_city}
\alias{normalize_city}
\title{Normalize city names}
\usage{
normalize_city(x)
}
\arguments{
\item{df}{A dataframe with municipality column and muni_code column}

\item{city_col}{Name of the municipality column (default: "")}
}
\value{
Normalized municipal names in the dataframe
}
\description{
Normalizes city names, removing punctuation marks, fixing misspellings and abbreviations.
}
\examples{
\dontrun{
# Add parent municipality to a dataframe
df <- normalize_city(df, city_col = "")
}
}
